+++ example/layout/index.vue
@@ -7,6 +7,7 @@
       <router-link to="/task-tree" class="box">TASK-TREE</router-link>
       <router-link to="/medal" class="box">MEDAL</router-link>
       <router-link to="/tag" class="box">TAG</router-link>
+      <router-link to="/icon" class="box">ICON</router-link>
     </div>
     <div class="routeBox">
       <router-view />