+++ types/icon.d.ts
@@ -1,0 +1,5 @@
+import { VFUIComponent } from "./component";
+
+export declare class VfIcon extends VFUIComponent {
+  name: string;
+}